Assisted design (for pid of lockbox output) is now avaliable + user defined analog tf
Left to do:
- Scope state should save rolling/normal property (+ running
continuous/stopped)
- Lockbox should have a button "record transfer function" using a state
"lockbox_tf" of the na
- Implement autorelock, but before that we need to discuss how to
diagnose the lock/unlock. I would like to add an option to read many
consecutive samples with the Sampler, in order to draw nice histograms
on the x/y axis of the input graphs. Also, from the std/mean of these
histogram, it becomes very easy to diagnose a delock...