Adds synchronization feature for different IQ modules and fixes bug with RedPitaya standalone.
Any IQ module now has the method synchronize_iqs(), which will cause the output signals to be in phase if the frequencies are commensurate at the time of synchronization.
The bugfix for standalone RedPitaya object concenred the use of the Module.pyrpl property even if no Pyrpl instance is available.